requirements_simplified.txt
matplotlib
networkx
numpy
torch
torchvision
torch-geometric
scikit-learn
xgboost==3.0.0
pandas
qiskit-aer
qiskit-algorithms
qiskit-ibm-runtime
scipy
##qiskit==1.4.1
##Install torch CPU not cuda.
